Refactor: creating separate tests for each model
